refactored user projects view
